requirements.txt
openai==2.16.0
python-dotenv==1.2.1
streamlit==1.38.0
